requirements.txt
python-telegram-bot==20.4
g4f==0.5.7.3
cryptography==41.0.4
pysqlite3==0.4.8
requests==2.31.0
Pillow==10.0.1
beautifulsoup4==4.12.2
lxml==4.9.3